fix sysinfo, try 2. it seems to work this time.

+++ include/sys/sysinfo.h
@@ -10,15 +10,15 @@ extern "C" {
 struct sysinfo {
 	unsigned long uptime;
 	unsigned long loads[3];
-	unsigned long long totalram;
-	unsigned long long freeram;
-	unsigned long long sharedram;
-	unsigned long long bufferram;
-	unsigned long long totalswap;
-	unsigned long long freeswap;
+	unsigned long totalram;
+	unsigned long freeram;
+	unsigned long sharedram;
+	unsigned long bufferram;
+	unsigned long totalswap;
+	unsigned long freeswap;
 	unsigned short procs, pad;
-	unsigned long long totalhigh;
-	unsigned long long freehigh;
+	unsigned long totalhigh;
+	unsigned long freehigh;
 	unsigned mem_unit;
 	char __reserved[256];
 };
@@ -26,8 +26,8 @@ struct sysinfo {
 int sysinfo (struct sysinfo *);
 int get_nprocs_conf (void);
 int get_nprocs (void);
-long long get_phys_pages (void);
-long long get_avphys_pages (void);
+long get_phys_pages (void);
+long get_avphys_pages (void);
 
 #ifdef __cplusplus
 }
